fix: dirty width calculate bug

+++ source/include/sgl_core.h
@@ -234,9 +234,9 @@ typedef struct sgl_surf {
     sgl_color_t *buffer;
     int16_t      x;
     int16_t      y;
-    int16_t      w;
+    uint16_t     w;
     int16_t      h;
-    size_t       size;
+    uint32_t     size;
 } sgl_surf_t;
